requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2026.3
dj-database-url==3.1.2
Django==6.0.4
django-crispy-forms==2.6
djangorestframework==3.17.1
gunicorn==25.3.0
packaging==26.2
pillow==12.2.0
psycopg2-binary==2.9.12
sqlparse==0.5.5
whitenoise==6.12.0
